Add sharing modes
I've updated how we specify that a dependency can be reused to be more user oriented instead of exposing a potential implementation in the PEP. This introduces sharing modes: none and group, to accomplish the same things as matchInterface, matchNamespace and IgnoreLabels.

I have also added more implementation details from the POC and answered some open questions about compatibility.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>